Scripts/Global : Finalize "Zero Warning Build"
- Final batch of stability and style fixes across various zones (Kalimdor, Draenor, Legion).
- Resolved signed/unsigned mismatches (int32 vs size_t) in warden_prison.cpp.
- Added missing f suffixes to double literals in broken_shore.cpp and shields_rest.cpp.
- Standardized frand() and M_PI logic across 15+ boss files including Siamat, Yalnu, and the Demonic Inquisition.
- No logic changes; focused purely on type safety and compiler compliance.

+++ src/server/scripts/Legion/warden_prison.cpp
@@ -29,7 +29,7 @@ public:
             if (!isUse)
                 return true;
 
-            for (int32 i = 0; i < sizeof(q38690); ++i)
+            for (size_t i = 0; i < sizeof(q38690); ++i)
             {
                 if (!player->GetReqKillOrCastCurrentCount(QUEST, q38690[i]))
                 {
